==> workspace/notebooks/04d_DEgenes_WTvsDEAB.Rmd <==
---
title: "Integrated HB13hpf DEAB_WT DEgenes WT vs DEAB R Notebook"
output: 
  html_notebook:
    toc: TRUE
    toc_depth: 3
    toc_float: TRUE
---

# 1. libraries and functions
```{r, results=F}
suppressPackageStartupMessages({
  library(Seurat)
  library(org.Dr.eg.db)
  library(BSgenome.Drerio.UCSC.danRer11)
  library(Signac)
  library(knitr)
  library(kableExtra)
  library(dplyr)
  library(ggplot2)
  library(ggsci)
  library(EnhancedVolcano)
})
options(future.globals.maxSize = 4000 * 1024^2)
```

```{r}
mypal <- pal_igv(palette = "default",alpha = 1)(50)
```

# 2. Read data

```{r}
neural <- readRDS(file = "RDSfiles/HB13.WT.DEAB.neural.refMap.int.RDS")
```

```{r}
DimPlot(neural, reduction = "ref.umap", group.by =  "predicted.Clusters", split.by = "orig.ident") + scale_color_igv()
```


```{r, fig.width=5, fig.height=5}
DefaultAssay(neural) <- "SCT"
for(cluster in levels(neural)){
  try({
    neural.subset <- subset(neural, idents = cluster)
    Idents(neural.subset) <- "orig.ident"
    markers <- FindMarkers(neural.subset, ident.1 = "WT", ident.2 = "DEAB", recorrect_umi = FALSE)
    write.table(markers, file = paste0("../results/WTvsDEAB/WTvsDEAB_",cluster,"_markers.txt"), sep = "\t", quote = FALSE, col.names = NA)
  })
}
```

```{r, fig.width=18, fig.height=15}
DefaultAssay(neural) <- "SCT"
plotlist <- list()
markerslist <- list()
for(cluster in c("r1","r1 & r2","r2","r3","r4")){
  try({
    neural.subset <- subset(neural, idents = cluster)
    Idents(neural.subset) <- "orig.ident"
    markerslist[[cluster]] <- FindMarkers(neural.subset, ident.1 = "DEAB", ident.2 = "WT", recorrect_umi = FALSE)
    p <- EnhancedVolcano(markerslist[[cluster]],
                         lab = rownames(markerslist[[cluster]]),
                         x = "avg_log2FC",
                         y = "p_val_adj",
                         title = paste(cluster, "Volcano plot"))
    plotlist[[cluster]] <- p
  })
}
vol_plots <- wrap_plots(plotlist)
vol_plots
ggsave(filename = "../results/FigS5_DEABvsWT_volcano.png",plot = vol_plots)
```

```{r}
sessionInfo()
```
